add users list page for admins

+++ kyberkucharka-backend/src/routes/userRoutes.ts
@@ -5,6 +5,7 @@ import {
   getPartialRecipesByUser,
   getUserByUsername,
   getUsers,
+  toggleAdminForUser,
 } from "../databaseFunctions";
 import { authenticateToken } from "../auth";
 
@@ -38,6 +39,13 @@ router.delete(
   (req: Request, res: Response) => {
     const username = req.params.username;
 
+    if (username === "admin") {
+      res.status(403).json({
+        message: `User "admin" cannot be deleted.`,
+      });
+      return;
+    }
+
     if (res.locals.user.username === username || res.locals.user.is_admin) {
       deleteUser(req.params.username ?? "")
         .then(() => {
@@ -55,6 +63,39 @@ router.delete(
         message: "You don't have the permission to delete this user.",
       });
     }
+  }
+);
+
+// update admin status for user
+router.put(
+  "/modify-admin/:username",
+  authenticateToken,
+  (req: Request, res: Response) => {
+    const username = req.params.username;
+    const should_be_admin = req.body.should_be_admin;
+
+    if (username === "admin") {
+      res.status(403).json({
+        message: `User "admin" cannot have their admin rights edited (duh!)`,
+      });
+      return;
+    }
+
+    if (!res.locals.user.is_admin) {
+      res.status(403).json({
+        message:
+          "You don't have the permission to update the admin status of this user.",
+      });
+      return;
+    }
+
+    toggleAdminForUser(username ?? "", should_be_admin)
+      .then(() => {
+        res.status(200).json({});
+      })
+      .catch(() => {
+        res.status(404).json({ message: "User not found" });
+      });
   }
 );
